Reorganize psp definitions into one file
